// src/Types/CitySearch.ts
export interface City {
  id: number;
  name: string;
  country: string;
  country_code: string;
  country_id: number;
  admin1: string;
  admin1_id: number;
  admin2: string;
  admin2_id: number;
  elevation: number;
  feature_code: string;
  latitude: number;
  longitude: number;
  population: number;
  timezone: string;
}